// src/enums/OrderStatusEnum.ts
export enum OrderStatusNumEnum {
  ALL = -1,
  PENDING = 0,
  CONFIRMED = 1,
  SHIPPING = 2,
  DELIVERED = 3,
  CANCELLED = 4
}

export enum OrderStatusTextEnum {
  ALL = 'ALL',
  PENDING = 'PENDING',
  CONFIRMED = 'CONFIRMED',
  SHIPPING = 'SHIPPING',
  DELIVERED = 'DELIVERED',
  CANCELLED = 'CANCELLED'
}